ref(searchQueryBuilder): Move GetTagValues to searchQueryBuilder
