+++ static/app/views/detectors/datasetConfig/components/releaseSearchBar.tsx
@@ -1,5 +1,6 @@
 import {fetchTagValues} from 'sentry/actionCreators/tags';
 import {SearchQueryBuilder} from 'sentry/components/searchQueryBuilder';
+import type {GetTagValues} from 'sentry/components/searchQueryBuilder';
 import type {FilterKeySection} from 'sentry/components/searchQueryBuilder/types';
 import {defaultConfig, InvalidReason} from 'sentry/components/searchSyntax/parser';
 import {t} from 'sentry/locale';
@@ -7,7 +8,6 @@ import type {TagCollection} from 'sentry/types/group';
 import {SavedSearchType} from 'sentry/types/group';
 import useApi from 'sentry/utils/useApi';
 import useOrganization from 'sentry/utils/useOrganization';
-import type {GetTagValues} from 'sentry/views/dashboards/datasetConfig/base';
 import {
   SESSION_STATUSES,
   SESSIONS_FILTER_TAGS,
