+++ static/app/views/explore/hooks/useGetTraceItemAttributeValues.tsx
@@ -1,6 +1,7 @@
 import {useCallback} from 'react';
 
 import {normalizeDateTimeParams} from 'sentry/components/organizations/pageFilters/parse';
+import type {GetTagValues} from 'sentry/components/searchQueryBuilder';
 import type {PageFilters} from 'sentry/types/core';
 import {defined} from 'sentry/utils';
 import {FieldKind} from 'sentry/utils/fields';
@@ -8,7 +9,6 @@ import type {ApiQueryKey} from 'sentry/utils/queryClient';
 import useApi from 'sentry/utils/useApi';
 import useOrganization from 'sentry/utils/useOrganization';
 import usePageFilters from 'sentry/utils/usePageFilters';
-import type {GetTagValues} from 'sentry/views/dashboards/datasetConfig/base';
 import type {
   TraceItemDataset,
   UseTraceItemAttributeBaseProps,
